Refactor : 올바른 가격 입력인지 확인하는 메소드 priceCheck 클래스로 변경
