+++ src/main/java/lotto/model/Price.java
@@ -3,7 +3,6 @@ package lotto.model;
 public class Price {
     private final int price;
 
-
     public Price(int price) {
         this.price = price;
     }
